Move stage components into new  directory
