+++ lit/src/components/modules/chat/chat_config.scss
@@ -1,4 +1,4 @@
-@use "../../sass/common";
+@use "../../../sass/common";
 
 :host {
   @include common.flex-column;
