+++ lit/src/components/modules/info/info_config.ts
@@ -1,12 +1,12 @@
-import "../../pair-components/textarea";
+import "../../../pair-components/textarea";
 
 import { observable } from "mobx";
 import { MobxLitElement } from "@adobe/lit-mobx";
 import { CSSResultGroup, html } from "lit";
 import { customElement, property } from "lit/decorators.js";
 
-import { core } from "../../core/core";
-import { ExperimentService } from "../../services/experiment_service";
+import { core } from "../../../core/core";
+import { ExperimentService } from "../../../services/experiment_service";
 
 import { styles } from "./info_config.scss";
